requirements.txt
pip~=23.1
wheel~=0.38.4
cryptography~=40.0.2
pytz~=2023.3
Babel~=2.12.1
setuptools~=65.5.1
Django~=4.2
asgiref~=3.6.0
sqlparse~=0.4.3
py-moneyed~=2.0
djangorestframework~=3.14.0
django-crispy-forms~=2.0
cffi~=1.15.1
MarkupSafe~=2.1.2
Werkzeug~=2.2.3
pycparser~=2.21